test(integration): Refactor nest-application to be generic
- Refactors the nest-application integration tests
so more tests can be implemented, since they
were specific to "getUrl"
- Adds test cases for X-Powered-By
